requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
 